+++ vector/src/main/java/im/vector/app/features/createdirect/CreateDirectRoomActivity.kt
@@ -91,7 +91,8 @@ class CreateDirectRoomActivity : SimpleFragmentActivity() {
                     KnownUsersFragment::class.java,
                     KnownUsersFragmentArgs(
                             title = getString(R.string.fab_menu_create_chat),
-                            menuResId = R.menu.vector_create_direct_room
+                            menuResId = R.menu.vector_create_direct_room,
+                            isCreatingRoom = true
                     )
             )
         }
@@ -121,7 +122,10 @@ class CreateDirectRoomActivity : SimpleFragmentActivity() {
 
     private fun onMenuItemSelected(action: UserDirectorySharedAction.OnMenuItemSelected) {
         if (action.itemId == R.id.action_create_direct_room) {
-            viewModel.handle(CreateDirectRoomAction.CreateRoomAndInviteSelectedUsers(action.invitees))
+            viewModel.handle(CreateDirectRoomAction.CreateRoomAndInviteSelectedUsers(
+                    action.invitees,
+                    action.existingDmRoomId
+            ))
         }
     }